backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.4.0
pymongo>=4.7.0
pydantic[email]>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
openai>=1.35.0
python-dotenv>=1.0.0
